api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
billiard==4.2.0
boltons==23.1.1
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cohere==4.45
colorama==0.4.6
colored==2.2.4
coloredlogs==15.0.1
cryptography==42.0.3
dataclasses-json==0.6.4
Deprecated==1.2.14
distro==1.9.0
eliot==1.15.0
eliot-tree==21.0.0
exceptiongroup==1.2.0
fastavro==1.9.3
Flask==2.2.2
frozenlist==1.4.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
griffe==0.36.9
grpcio==1.60.1
guardrails-ai==0.4.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.5
langchain-core==0.1.23
langsmith==0.0.87
libmagic==1.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.3
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
orjson==3.9.14
packaging==23.2
pdfminer==20191125
pdfminer.six==20231228
prompt-toolkit==3.0.43
protobuf==4.25.3
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.4.2
pydantic_core==2.10.1
pydash==7.0.7
Pygments==2.17.2
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
PyPDF2==3.0.1
pyreadline3==3.4.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rstr==3.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
toolz==0.12.1
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.2
win_unicode_console==0.5
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2
